patch 8.0.0387: compiler warnings
Problem:    compiler warnings
Solution:   Add type casts. (Christian Brabandt)

+++ src/channel.c
@@ -1966,7 +1966,7 @@ channel_parse_json(channel_T *channel, ch_part_T part)
 	     * more (but still incomplete): set a deadline of 100 msec. */
 	    ch_logn(channel,
 		    "Incomplete message (%d bytes) - wait 100 msec for more",
-		    buflen);
+		    (int)buflen);
 	    reader.js_used = 0;
 	    chanpart->ch_wait_len = buflen;
 #ifdef WIN32
@@ -3299,6 +3299,7 @@ channel_read_block(channel_T *channel, ch_part_T part, int timeout)
 	channel_read(channel, part, "channel_read_block");
     }
 
+    /* We have a complete message now. */
     if (mode == MODE_RAW)
     {
 	msg = channel_get_all(channel, part);
